requirements.txt
# === Core ===
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings>=2.0.0
python-dotenv==1.0.1
cryptography>=42.0.1
PyJWT==2.9.0
httpx>=0.27.0

# === Database ===
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pgvector>=0.3.0

# === Redis & Task Queue ===
redis>=5.0.0
celery[redis]>=5.3.0

# === AI / LLM ===
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.3.0
langgraph>=0.2.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# === Utilities ===
faker==30.8.2

# === UI (optional — only for Streamlit chat) ===
streamlit>=1.40.0